adding remedy_utils.get_ip()
